backend/functions/requirements.txt
fastapi
uvicorn
requests
google-cloud-firestore
functions-framework
python-dotenv
sendgrid
